Add board-cfg for PCoreMX8MN-FERT20
The timings for the Insignis NDL46PFP-8KIT (B.ICDRAM.21) were generated
in the old mscale DDR tool from NXP. The new Config Tool has some issues
for the i.MX8M Nano.
For the Insignis DRAM the second frequency setpoint is missing on
purpose, because with it the UBoot could not be started. It seems to be
an issue specific to the Insignis DRAM, as a PCoreMX8MN-FERT5 could be
started with the same timings and two frequency setpoints.