Actually skip full bad blocks in nand_util.c
When writing to a page that is not at the beginning of a block (or even
at the beginning of an MTD partition), it means the caller wants to
continue writing to an already partly written image. So it is essential
that the page offset is kept if the first block is bad. So if writing
to page n in block x, and block x is bad, skip block x and write to page
n in block x+1. The previous code wrote to page 0 of block x+1, which
does not make sense.

The same is true when reading data.

So change functions check_skip_len(), nand_write_skip_bad() and
nand_read_skip_bad() accordingly.

If reading or writing starts with the first page of a block, for
example when writing to a partition, the result is the same as before,
so in typical cases, this should make no difference. But we need this
new behaviour when reading and writing images with fsimage load/save.

Also fix an indentation error in nand_base.c.