Add support to write SPL to NAND
Add a temporary buffer temp to struct flash_info to hold one NAND page.
When writing data in fs_image_save_region_to_nand(), use new function
fs_image_save_sub_to_nand() to actually write a sub-image. This function
uses the temp buffer when handling image parts that are not aligned to or
smaller than a page. Add function fs_image_flush_temp_nand() that
saves the temp buffer and function fs_image_write_to_nand() that saves
a page-aligned part of the image. These functions also handle existing
bad blocks, so that we can write larger images in parts.

Add flag SUB_SYNC to mark a sub-image, after that the temp buffer should
be written back. This is typically the last image of a region or before
a gap. For example BOARD-CFG currently gets the SUB_SYNC flag because
there is a gap to the FIRMWARE sub-image behind it and the FIRMWARE will
definitely start on a new page boundary.

In fs_image_save_nboot_to_nand(), add code to setup FCB, DBBT and
DBBT-DATA and prepare an additional BCB region that is also saved in
addition to the SPL and NBOOT regions.

Have a function fs_image_put_flash_info() that is the counterpart to
fs_image_get_flash_info(). It is called at the end of an fsimage load or
save command and frees all allocated resources like the temp buffer. In
case of eMMC, we can use this place to switch back to the original hwpart,
so we can get rid of function fs_image_set_old_hwpart().

Add a big comment at the beginning of the file that explains the
relationship of regions, storage-infos and sub-images.