On fsimage save, do not fail if at least one copy could be saved
If at least one copy could be saved completely when doing fsimage save,
do not return with an error, only issue a warning about an unstable
system. This one copy should be enough to start the system, but of
course there is no safety margin anymore and any additional problem can
cause the system to fail.