Add ops to struct flash_info in fsimage.c
To handle flash access (loading and saving) as generic as possible, there
is a struct flash_info that holds all flash specific settings. Add an ops
entry of type struct flash_ops that holds a set of access functions that
work differently in case of NAND and eMMC. When the common code requires
a flash specific procedure, it calls such an ops function. As a result,
the remaining code can be kept common for both types of flash memory.

fs_image_save_region() and fs_image_flush_temp() are now generic versions
that can be used for NAND and eMMC, we just need a small flash specific
write function. Remove block number output for eMMC, this does not make
sense when writing smaller entities like a BOARD-ID or a single F&S header
in the future anyway, and we do not need different functions just for
presenting this info.

Add function fs_image_check_crc32() that verifies the CRC32. This
simplifies the code for fs_image_validate() and can also be used when
loading images.

On fsimage load, also load and verify Secondary Image Table in case of
eMMC and when loading from User area.

Sort functions that they are in the sequence of struct flash_ops.

Add debug output for nboot_info.