Add fsimage load uboot to load U-Boot image
Command fsimage load used to load NBoot. This can be used to check if all
copies of NBoot are still functioning. Loading will show warnings if not,
and then the loaded NBoot image can be saved again with fsimaage save to
refresh the copies.

The same should be possible with U-Boot. So add optional argument
"uboot" to fsimage load to allow loading U-Boot instead of NBoot. For
symmetrical reasons, also accept argument "nboot" to load NBoot, but
of course this remains the default.

However there is a problem to detect the size of U-Boot. U-Boot is stored
without an F&S header and it is rather complicated to determine the size
of a FIT image without that. Add function fs_image_get_size_from_fit()
that does exactly that. It scans all images of the /images node and
determines offset and size. The end of that specific external image with
the highest offset is also the end of the FIT image, i.e. U-Boot. Call
the new function from fs_image_load_nand()/fs_image_load_mmc() if the
magic number shows an FDT header.

Add function fs_image_load_uboot() to actually load U-Boot. Call from
do_fsimage_load() if argument "uboot" is given.

Fix fs_image_load_image() to load images aligned to 16 bytes by setting
the pad entry in the F&S header correctly. Up to now, all NBoot images
were correctly aligned already, but the U-Boot image may not be aligned.

Add an ops function read() to read data from NAND or eMMC. Similar to the
write() function, this gets uint arguments and only uses NAND and MMC
specific types inside. Call this function wherever a direct blk_dread()
was used in the past.

In a previous commit, we had moved the information output, what is
currently loaded, to a later point when the size of the file is known.
Then it was possible to show the size in parentheses exactly in the same
way as when saving. However this had the unwanted side effect that if the
error happens in the part where the size itself is determined, that no info
was output yet and the FAILED message was appearing without any context.
So move the size for all output (loading and saving) to the end of the
info string and in case of loading, add it slightly later when it is known.
Then we have at least some basic information available if the error happens
before the size is known.