Store real BOARD-ID in BOARD-CFG
In the past, the real BOARD-ID was not stored because originally the ID
of the finally used BOARD-CFG should be identical to the BOARD-ID given
to the board when installing the software for the first time. So we could
take the ID of the saved BOARD-CFG as the current BOARD-ID.

However we found out that not every minor board revision needs a different
software. This would increase the number of BOARD-CFG files considerably.
So we decided to use a fuzzier comparison. The newest BOARD-CFG, that is
not newer than the BOARD-ID, is taken. For example if there are BOARD-CFGs
for board revisions 110, 120 and 130, and the BOARD-ID is for revision 122,
then the BOARD-CFG for 120 is taken, as it is the newest one not newer
than the BOARD-ID.

However this meant that from now on, the board reported a BOARD-ID with
revision 120 because this was the ID of the savend BOARD-CFG. So it
"forgot" that it was originally a board-revision 122. In most cases, this
should not matter. Of course the output is silghtly incorrect, but other
than that, the software should behave identical.

But what if we find out later, that we actually do need a different
software for this board revision 122? Then we have a problem. We cannot
tell the difference anymore. So it is actually a good idea to store the
real BOARD-ID, too.

To avoid a completely new layout, we use a trick and store the BOARD-ID
in the BOARD-CFG, too. The ID of the BOARD-CFG can only differ in the
board revision, here 122 instead of 120. So it is sufficient to store
the different revision. The base name can be taken from the ID of the
BOARD-CFG. We store the revision in file_size_high of the F&S header,
which is always zero. So when a new NBoot is stored with fsimage save,
file_size_high is set to the real board revision (as a number), which
is 122 (0x7a) in our example. The value is typically between 100 and
999. Even if we assume, that four or five digits might be used in the
future, the value should still fit into 16 bits.

SPL will load this BOARD-CFG and stores it in OCRAM, where it will be
used by U-Boot. So the BOARD-CFG in OCRAM will also hold the BOARD-ID
coded in file_size_high.

There are two things to consider. First, changing anything in the
BOARD-CFG will corrupt its signature in a Secure Boot environment. So
we have to set back file_size_high to 0 before authenticating the
BOARD-CFG. And second, this BOARD-ID value itself can be corrupted,
even if the signature without it is OK. So we add a CRC32 to cover
the BOARD-ID, too. We have a CRC32 in the non-secure version anyway,
so we only have to update it accordingly.

fs_image_common.c:

- Move fs_image_check_crc32() from fsimage.c to fs_image_common.c, it
  is needed in SPL, too.
- Change fs_image_cfg_is_valid() to fs_image_is_ocram_cfg_valid(). It
  will check the CRC32 first, then, in case of Secure Boot, it undoes
  all modifications, performs the regular signature authentication and
  applies the modifications again. Call it in fs_image_load_system()
  after the BOARD-CFG is loaded.
- Add function fs_image_board_cfg_set_board_rev() that sets the board
  revision of the BOARD-ID to the BOARD-CFG in OCRAM if a new one
  was loaded (e.g. when downloading a new NBoot with SDP/UUU) and
  update the CRC32.
- Change the board-id and compare_id handling accordingly.
- Drop argument "type" from fs_image_match_board_id(), it has to be
  "BOARD-CFG", there is no choice.

fsimage.c:

- When saving NBoot, also set the board revision and update the CRC32
  of the BOARD-CFG sub-image by calling fs_image_board_cfg_set_board_rev().
- When loading NBoot, remove the board revision of the BOARD-CFG again and
  update the CRC32 accordingly.
- Call fs_image_set_board_id_from_cfg() in do_fsimage(), so that the values
  are up to date.
- Change fs_image_find_board_cfg() to use the new id handling.
- do_fsimage_boardid() is simpler now because of the new id handling.

TODO:
The Secure Boot case is not fully done yet and should not be used.