Prepare for Secondary SPL detection
When the Primary SPL is damaged, the ROM Loader loads the Secondary SPL.
In this case, we should also load NBoot (BOARD-CFG, FIRMWARE) and
probably even U-Boot from the secondary copy. Mixing copies may result in
non-matching images, e.g. if only one copy was updated to a newer version.
If old and new version are incompatible, then booting may even fail in
such cases.

We should also update the "other" copy first, when writing NBoot and
U-Boot with fsimage save. So we also need to pass this information from
SPL to U-Boot. This is done by setting the most significant byte (MSB) of
file_size_high in the F&S header of the BOARD-CFG in OCRAM different to
non-zero if running from the Secondary SPL. Please note that file_size_high
is also used for encoding the BOARD-ID. But the BOARD-ID only uses no more
than the lower 16 bits of file_size_high, so the MSB is still zero.

Of course setting this byte also corrupts the CRC32 and the signature of
the BOARD-CFG in OCRAM, so we have to clear this information rather early
in U-Boot again before doing any validation of the BOARD-CFG.

- Add function fs_image_mark_secondary() that marks the BOARD-CFG in OCRAM
  to tell that we were booting from Secondary SPL. Add a call to it in
  board_init_f() if we are actually on the Secondary SPL.
- Add function fs_image_is_secondary() that checks the mark in BOARD-CFG
  in OCRAM and removes it. Call this function in fs_setup_cfg_info() in
  the board file fsimx8{mm,mn,mp,x}.c, add new flags entry to cfg_info and
  add flag CI_FLAGS_SECONDARY that indicates running on the Secondary copy.
- Check this flag in fs_image_save_nboot_nand(), fs_image_save_nboot_mmc()
  and fs_image_save_uboot() and write "the other" copy first, i.e. the copy
  that we are not currently running on. The idea is that if the other copy
  is probably damaged, we should try to write a working version there first
  before touching the still working copy.

Add function is_imx8m_running_secondary_boot_image() to soc.c to detect if
the system was booted from Primary or Secondary SPL. This function evaluates
the events that are recorded in the ROM Loader Boot-Log. If there is an event
0x51, we know that the system started from Secondary SPL. This is how NXP
detects the eMMC boot partition and they say this should work for all i.MX8Mx
CPU variants. However it doesn't. It only works on i.MX8MM, but not on
i.MX8MN/MP. So actually we still do not have a way to detect this reliably.

TODO:
Instead of relying on the CPU to tell us on which SPL we are running, we could
detect this ourselves. When writing the Primary SPL, we could set a specific
byte to 0 and when writing the Secondary SPL, we could set the byte to 1. Then
we could check the byte at runtime and we know which copy we are running.

However setting such a byte would corrupt the SPL signature. So we have to
find a byte that is loaded by the ROM loader on one hand, but is not covered
by the Secure Boot Signature on the other hand. This requires some deeper
inspection of the CSF and the Secure Boot Signing process, so we postpone
this topic for now until Secure Boot is working again.