Add NBoot capability properties to nboot-info
The original idea was to use the board_cfg_size to detect whether we have
the new nboot-info layout with many new features or the old layout without
these features. Thinking longer about it, this might not be such a good
idea. We have added many new features that are still untested in the field.
Maybe we have to remove or change a single feature again, then it is
impossible to detect this from the board_cfg_size alone. And if we ever
add new feature in the future, we stand before the same problem again.

A better idea is to actually tell which things are supported. Of course
some things are easy to detect. For example if there is an env-start entry
in nboot-info, then we obviously have the env location given in nboot-info
and not in the device tree anymore. But for things that are not easy to
decide, add an own boolean property. The following properties are available
now:

support-crc32
-------------
NBoot will handle CRC32 settings in F&S headers correctly. This means image
type BOARD-CONFIGS is expected as BOARD-INFO and image type DRAM-SETTINGS is
expected as DRAM-INFO now. (The old entries were too long and would collide
with the CRC32.)

save-board-id
-------------
NBoot is capable of extracting the BOARD-ID (or more precisely the real
board revision) from the stored BOARD-CFG in file_size_high of the F&S
header. It will also set it in the BOARD-CFG in OCRAM in the same way.

emmc-both-bootparts
-------------------
If booting from eMMC partition boot1 or boot2, then both partitions are
used, one copy in boot1, one copy in boot2, with identical layout. On
fsimx8mm, both copies of SPL, BOARD-CFG and FIRMWARE used to be in the
same boot partition. This will change in the future, so we have this
property already. Then fsimx8mm will be more compatible to the other
architectures, where this beaviour has been the default all the time.
You should nonetheless set this property in new nboot-infos even for
these architectures to make this behaviour visible.

uboot-emmc-bootpart
-------------------
If booting from eMMC partition boot1 or boot2, then also U-Boot is stored
in the boot partitions. Because only one U-Boot copy will fit into a boot
partition, you also have to set emmc-both-bootparts in this case.

uboot-with-fsh
--------------
NBoot (i.e. SPL) is capable of loading an U-Boot image with a prepended
F&S header. This allows giving U-Boot a CRC32, too.