Add new nboot-info layout scheme in nboot-info.dtsi
Add new layout scheme for F&S i.MX8 boards. Instead of having two separate
files nboot-info-mmc.dtsi and nboot-info-nand.dtsi, there is only one file
nboot-info.dtsi now. It contains the layout when booting from NAND, the
layout when booting from an eMMC boot partition and the layout when booting
from the eMMC USer partition or from an SD Card. These layouts are provided
as sub-nodes of the nboot-info node with the names "nand", "emmc-boot" and
"sd-user".

At the same time simplify the nboot-info build process.

- The BOARD-CFG files are now preprocessed and can use macros. We include
  config.h in nboot-info.dtsi, so all CONFIG macros are available.
  CONFIG_SPL_BUILD is also set, so the files logically belong to SPL.
  This avoids having to add some macro names to config_whitelist.txt.
- The properties board-cfg and board-rev are now automatically set. They
  can be determined automatically in the build process, so it is not
  necessary to set them in each BOARD-CFG file.

For these changes all BOARD-CFG files needed to be touched. Update nboot
version to 2023.08 on all platforms. This is the turning point where the
new features are supported.

- Fix a wrong nboot-start address on fsimx8mp. NBOOT should be stored at
  offset 0x40000, not 0x80000. The old value could overwrite the U-Boot
  environment if the files grew bigger.
- In fs_image_common.c, the new sub-node needs to be considered in
  fs_image_loop().
- In the board spl.c files, function basic_init() determines the U-Boot
  offset. This is requested via spl_mmc_get_uboot_raw_sector() from
  the U-Boot load infrastructure. To avoid having to do flash specific
  stuff there, fs_image_common.c passes the name of the layout sub-node to
  basic_init() now.
- The new sub-node also has to be considered in fs_mmc_get_env_info()
  (in fs_mmc_common.c) and in fs_nand_get_env_info() (in the board files
  that support NAND).
- The new sub-node also has to be checked in fs_image_get_nboot_info_nand()
  and fs_image_get_nboot_info_mmc() in fsimage.c.