Have new NAND/MMC layouts for fsimx8mm/mn/mp
The layouts for MMC up to now were mixed. SPL and NBOOT were in the boot
partitions, the environment also, but slightly different and U-Boot was
stored in the User partition. With the new layout scheme, we can now have
two clean setups: one to store everything in the boot partitions, and one
to store everything in the User area. Which one is used depends on the
boot partition set in the eMMC itself. This can be changed with command
mmc partconf. By default, we should configure to boot from boot1.

As we can write SPL now directly from U-Boot, we can also optimize the
layout.

These are the ideas behind the new layouts:

- Have settings as similar as possible for emmc-boot and sd-user.
- Put NBOOT next to U-BOOT. If either one grows in the future, we can
  shrink the other without needing to rearrange everything else.
- Increase the size of NBoot, because ATF and opTEE may get bigger and
  bigger in the future.
- In the NAND layout, have enough reserve for Bad Blocks everywhere.
- If there are unused areas, move them next to the environment. We do
  not know if we need more than 16 KiB in the future.

Also change the environment addresses in the defconfigs. However as
the positions differ between emmc-boot and sd-user, we can only store
one setting. We use the sd-user setting, because in the emmc-boot
setting the locations for normal and redundant environments are always
the same, just in the other hwpart. This may confuse U-Boots internal
logic.

As the NAND and MMC layout is now in the nboot-info.dtsi, remove them
from the head comment in fsimx8{mm,mn,mp}.h files.

TODO: Add a similar layout for fsimx8x, too.