Improve fsimx8mm device trees
Use linux device trees and add U-Boot specific changes in the
-u-boot.dts files.

Use a common picocoremx8mm-u-bootinc-dtsi file for all variants to reduce
redundancy.

Some changes hat to be made to the linux device trees, like adapting the
path to the imx8mm.dtsi or removing some nodes that are not present in
the uboot imx8mm.dtsi.

Also add the nxp,pfc85263.h so we dont get include errors.