Fix bootdev initialization with unfused usb boot
In case of an unfused USB boot the bootdev did not get initialized.
tehrefore the nand device was also not initialized in Uboot which lead
to an cpu hang when trying to read the NAND size.

Simplify the bootdev initialization and return an error in case of usb
boot so the bootdev is initialized from the config from usb.

Also remove the "start" variable and always initialize the bootdev the
first time.  The "boot_dev_init_done" variable should prevent that the
bootdev gets initialied muliple times.