Fix fsimage save when U-Boot is relocated
If U-Boot or environment needs to be relocated when saving a new NBoot
with fsimage save, the temp buffer is filled after reading the old file.
This caused fs_image_save_region() to flush that buffer, resulting in a
write to the old (wrong) offset. Fix this by dropping the temp buffer at
the beginning of fs_image_save_region().