Add command fsimage boot to show current boot settings
Add new subcommand boot to command fsimage. This command shows the
current boot settings, i.e. the boot device, the boot partition and
the places where alle images are located.

The implementation is rather simple because all this information is
already available in functions fs_image_get_nboot_info_nand() and
fs_image_get_nboot_info_mmc() as debug info. So add an additional
argument that tells these functions whether to show this info when
called or not and change the debug output to real output. Call with
true in the new function do_fsimage_boot() and with false in all other
cases.

Add subcommand to command table and add help text.