On fsimx8mm, add old device tree names for compatibility
Old BOARD-CFG variants used the name PicoCoreMX8MM for PicoCoreMX8MM-LPDDR4
and PicoCoreMX8MX for PicoCoreMX8MM-DDR3L. These names are used to determine
the dtb device tree name for U-Boot.

For updating from old NBoot/U-Boot combinations to new combinations, we
need to update U-Boot first and then restart, because only a new U-Boot
knows how to store a new NBoot image with all the new NBoot features.
However if the new U-Boot only has dtb device tree files with the new
names appended, then the old NBoot with the old dtb names in its
BOARD-CFGs will not find the dtb in the new U-Boot and fails to start
U-Boot.

So for now, also add copies of the device trees under the old names.
This increases the U-Boot size, but there is no other way. After a while
we can drop the old names again.

- Create symbolic links picocoremx8mm.dts and picocoremx8mx.dts that
  point to picocremx8mm-lpddr4.dts and picocoremx8mm-ddr3l.dts
  respectively.
- Add the old filenames to arch/arm/dts/Makefile.
- In fsimx8mm_defconfig, also include the old filenames in the list of
  added dtb files.