fsimx8mm/mn: Fix default entry for environment in NAND
If the environment is not given in NBoot and the legacy entry also
cannot be used, function fs_nand_get_env_info() falls back to the
CONFIG entries. Unfortunately, the redundand entry was also stored
at index 0 instead of index 1. Fix this in fsimx8mm.c and fsimx8mn.c.

While at it, also fix a bad size in a comment in nboot-info.dtsi for
fsimx8mm (MiB -> KiB).