addfsheader.sh: Check for crc32 and xxd before using them
If the crc32 command is not available on the system, addfsheader.sh
unintentionally uses an empty string and creates a bad CRC32 in type[12..15].
It also shifts the remaining part four bytes to the back, resulting in a bad
description with "\x\x" prepended.

So if CRC32 is requested, add a check for the crc32 command and fail if not
available. Also include this information in the usage text. On Fedora, crc32
is part of the perl-Archive-Zip package.

Command xxd is also not part of the regular command set as given by package
coreutils. On Fedora, it needs the additional package vim-common. Add a
check for xxd, too, and fail if not available.

While at it, also fix a comment, that still assumed the CRC32 value to be
somewhere else.