Add new USB handling in SPL and UBoot
ATTENTION: This implementation is not yet tested for the other fsimx8mp
boards! This is a prototype for the next release.
In UBoot the USB host and otg device will be determined from the
device-tree. This implementation moves board-specific code into the
device-tree to make the USB functions more readable.