Fix flashing process when USB device is not clear
NXP reads from the ROM pointer in addition to the boot_type the instance
for the boot device. Add USB2_BOOT in the handling of the boot_dev. The
changes only apply to imx8mn and imx8mp.
ATTENTION: In the future, this will also apply to imx8ulp and imx93, so
be aware to test it in future releases.