Add loading of secondary UBoot for fsimx8mn
The loading of the secondary UBoot was not implemented yet. Change the
loading process to fall back to the secondary UBoot when it fails to
load the primary.

In board/F+Sfsimx8mn/spl.c variables are added that hold the offset and
partition (for emnmc), of the secondary UBoot. Another variable called
uboot_try is added to keep track of the current Uboot that should be
loaded.

Functions that return the partition and offset use the uboot_try
variable to return the right partition and offset. The offset function
increases the uboot_try vartiable by one, so that the next iteration
starts with the right index.

The loader in common/spl/spl_mmc.c is changed to a new loader that calls
the original loader and checks if it is successfull. If it was not, it
will output a warning, mark the UBoot as secondary and try a second
time. The function call is the same, but since the uboot_try variable
was handled, it will load from another location.